Transferring the function to change the avatar moved to Image Repository and creating a singleton for FirebaseStorage and FirebaseFirestore
