+++ app/build.gradle
@@ -101,6 +101,7 @@ dependencies {
     implementation 'com.google.firebase:firebase-firestore-ktx:24.5.0'
     implementation 'com.google.firebase:firebase-config-ktx:21.3.0'
     implementation 'com.google.firebase:firebase-messaging-ktx:23.1.2'
+    implementation 'com.google.firebase:firebase-storage-ktx:20.2.0'
 
     implementation 'com.google.android.gms:play-services-maps:18.1.0'
