+++ domain/src/main/java/com/safetyheads/akademiaandroida/domain/repositories/UserRepository.kt
@@ -1,7 +1,5 @@
 package com.safetyheads.akademiaandroida.domain.repositories
 
-import android.graphics.Bitmap
-import android.net.Uri
 import com.safetyheads.akademiaandroida.domain.entities.ResetPassword
 import com.safetyheads.akademiaandroida.domain.entities.User
 import com.safetyheads.akademiaandroida.domain.entities.firebasefirestore.Profile
@@ -12,9 +10,4 @@ interface UserRepository {
     suspend fun resetPassword(email: String): Flow<ResetPassword>
     suspend fun getProfileInformation(userUUID: String): Flow<Result<Profile>>
     suspend fun logIn(email: String, password: String): Flow<Result<String>>
-    suspend fun addImageToFirebaseStorage(imageUri: Uri): Flow<Result<String>>
-    suspend fun addImageToFirebaseStorage(imageBitmap: Bitmap): Flow<Result<String>>
-    suspend fun addImageToFirebaseUserProfileFirestore(userUUID: String, imageStringReference: String): Flow<Result<String>>
-    suspend fun removeImageFromUserProfileFirestore(userUUID: String): Flow<Result<String>>
-    suspend fun removeImageFromFirebaseStorage(imageStringReference: String): Flow<Result<Boolean>>
 }
